app/backend/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.23
aiortc>=1.9
pyzmq>=25
numpy>=1.24
Pillow>=10.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
av>=10.0
httpx>=0.25
